Running the improved Fuzzy Cluster Method (FCMm) by optimizing the fuzzifier m

An improved version of Fuzzy Cluster Method (FCM) for water spectra data sets.

FCM.new(
  FDlist,
  K,
  sort.pos = length(FDlist$wv),
  sort.decreasing = FALSE,
  plot.jitter = TRUE,
  fast.mode = FALSE,
  do.stand = TRUE,
  stand = NULL,
  ...
)

Arguments

FDlist
A list from function FuzzifierDetermination

K
Number, cluster number

sort.pos
The position to sort the cluster number. Default as the end position of the input training matrix. It can be NULL if there is no need to re-sort the cluster index.

sort.decreasing
Default as FALSE which means the sort are assigned from small values to large values.

plot.jitter
Logical, choose to plot jitter plot using ggplot2 functions

fast.mode
Logical, FALSE (default)

do.stand
Whether to use standarized data for FCM. Default as TRUE

stand
Deprecated; Now stand = !do.stand

...
Parameters pass to fcm

Value

A list of FCM:

FD The return list by function FuzzifierDetermination

res.FCM The optimized FCM result generated by functions in package ppclust

K Cluster number

centroids The list of centroids (both at raw and normalized scale) of the cluster result by aggregating each mean.

plot.jitter A logical value for the option of doing jitter plot by package ggplot2

fast.mode A logical value for choosing whether to use fast mode

References

Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.

Dembele D. Multi-objective optimization for clustering 3-way gene expression data[J]. Advances in Data Analysis and Classification, 2008, 2(3): 211-225.

See also

Examples

library(FCMm) 
library(ggplot2) 
library(magrittr)
library(stringr)
data("Nechad2015")
x <- Nechad2015[,3:11]
wv <- gsub("X","",names(x)) %>% as.numeric
w <- sample(1:nrow(x), 100) 
x <- x[w, ]
names(x) <- wv
nb = 4 # Obtained from the vignette "Cluster a new dataset by FCMm"
FD <- FuzzifierDetermination(x, wv, stand=FALSE)

#> Warning: Parameter stand has been deprecated. Please use do.stand!

result <- FCM.new(FD, nb, fast.mode = TRUE)

#> The fast mode is used. Pay attention to the convergence of the obj. fun. and results!