Determine the optimized fuzzifier for FCM running

To determine the optimized fuzzifier value for FCM running.

FuzzifierDetermination(x, wv, max.m=10, stand=FALSE, dmetric='sqeuclidean')

Arguments

x
Data.frame. the input Rrs data

wv
Wavelength of X

max.m
Set max.m as for determination of m.mub. Default as 10

stand
Whether the data set is standardized (T needs not standardization)

dmetric
Distance method. Default as 'sqeuclidean'

Value

FD list contains several result by FuzzifierDetermination:

x The raw input Rrs dataframe with unit sr^-1

x.stand The standardized Rrs dataframe, if stand=F

wv Wavelength with unit nm

max.m The maximum fuzzifier of FCM as a restriction

stand A logic value for choosing whether to use standardization

dmetric A string value for choosing which distance metric to be used

Area A numeric vector for trapezoidal integral values

m.ub The upper boundary of fuzzifier(m) value

m.used The desired value of fuzzifier(m) value

References

Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.

Dembele D. Multi-objective optimization for clustering 3-way gene expression data[J]. Advances in Data Analysis and Classification, 2008, 2(3): 211-225.

See also

Examples

if (FALSE) {
library(FCMm)
library(tidyverse)
data("Nechad2015")
w <- Nechad2015 %>% names %>%
    str_extract(.,pattern="\\d") %>%
    is.na %>% {!.}
wv <- w %>% names(Nechad2015)[.] %>%
    gsub('X','',.) %>% as.numeric
x <- w %>% Nechad2015[,.]
names(x) <- wv
rm(w)
FD <- FuzzifierDetermination(x, wv, stand=F)
}