Get the result of function Assessment_via_cluster

This function mainly use function Assessment_via_cluster to get assesments both from fuzzy and hard mode. Specifically, it will return fuzzy_MAE, fuzzy_MAPE, hard_Slope, Valid_percent, and hard_Rsquare which should be seemed as the input value of function Scoring_system.

Getting_Asses_results(
  sample.size,
  replace = FALSE,
  pred,
  meas,
  memb,
  cluster,
  seed = NULL
)

Arguments

sample.size
Sample size. This supports a bootstrap way to run the function Assessment_via_cluster. The number should not be larger than the row number of pred or so.

replace
Logical, replace, default as FALSE

pred
Prediction matrix

meas
Measured (actual) vector

memb
Membership matrix

cluster
Cluster vector

seed
Seed number for fixing the random process. See help(set.seed) for more details.

Value

A list containing fuzzy and hard results from Assessment_via_cluster

Note

The row number of pred, meas, memb, and cluster should be the same. This function is designed for bootstrapping process to get Chla algorithms assessment. Therefore, parameters of Assessment_via_cluster is set as fixed such as log10 = TRUE, na.process = TRUE. Given that, I will not export this function in latter to avoid confuses.

See also