Apply the improved Fuzzy Cluster Method (FCMm) to new data based on the defined centroids

Application of the improved Fuzzy Cluster Method (FCMm) for new Rrs data based on default cluster settings or user-defined clusters (trained by FCM.new).

apply_FCM_m(
  Rrs,
  wavelength = NULL,
  Rrs_clusters = NULL,
  m_used = 1.36,
  do.stand = FALSE,
  stand = NULL,
  default.cluster = TRUE,
  quality_check = FALSE,
  option.plot = FALSE,
  color_palette = NULL
)

Arguments

Rrs
Data.frame, the input Rrs of FCM.

wavelength
Numeric vector, used for applying FCM. Default use the data from Bi_clusters.rda

Rrs_clusters
Data.frame, used for applying FCM. Default use the data from Bi_clusters.rda

m_used
Number, Used fuzzifier value

do.stand
Logical, whether to normalized the Rrs data (both for Input and Centroids). Default as FALSE means do not.

stand
Deprecated; same to do.stand.

default.cluster
Logical, whether to use the default clusters. Default use the data from Bi_clusters.rda

quality_check
Logical, quality chech option (default as FALSE)

option.plot
Logical, whether to plot the result. Default as FALSE

color_palette
The palette of cluster color. Default as RdYlBu(res$K) if NULL. It could be RdYlBu(k), Spectral(k), HUE(k) or other color values with same length of cluster number. k means the cluster number.

Value

A list including several results of function apply_FCM_m()

x The raw input Rrs dataframe with unit sr^-1

x.stand The standardized Rrs dataframe, if stand=F

d Distance to each cluster

u Membership values

Area Spectral intergration of each sample

cluster Defined by the maximum of membership

quality The quality of the cluster results.

m.used The used value of fuzzifier(m)

K Cluster number

p.group A ggplot list for plotting the cluster result

p.group.facet Recommened! p.group with facet to see each cluster results more clearly

dt.melt Dataframe used for ggplot including dt_plot_x, dt_plot_v, and cp.

References

Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.

See also

Examples

library(FCMm)
data("WaterSpec35")
data("Bi_clusters")
Rrs <- WaterSpec35[,3:17]
result <- apply_FCM_m(Rrs=Rrs, option.plot=TRUE)