apply_to_image

This function could apply the defined water cluster to corrected image files. Should run generate_param() to generate a res list as an input of function apply_to_image

apply_to_image(input, res,
  output_image=TRUE, output_resultpng=FALSE, output_imRrs.n=FALSE,
  Chla_est=FALSE,
  title.name = NULL, png_scale=50,
  fn_memb="output_membership",
  fn_cluster="output_cluster",
  fn_imRrs.n='output_imRrs_normalized',
  fn_truecolorpng='output_truecolor',
  fn_Chla = 'output_Chla',
  output_format='GTiff')

Arguments

input
A raster or a character linking to the raster file on the disk.

res
A required list that used for clustering the image data including:

FD The results from function FuzzifierDetermination providing wavelength

K Cluster number

res.FCM Cluster center and used fuzzifier value.

For the convenience, function generate_param() supports to quickly generate this list. See more in examples.

output_image
Logical, whether to produce image files

output_resultpng
Logical, whether to produce png files

output_imRrs.n
Logical, whether to produce normalized Rrs files

Chla_est
Logical, whether to estimate Chla concentration

title.name
Character, the title name of ggplot for plotting

png_scale
Numeric, scale of png

fn_memb
A string, filename of membership raster file

fn_cluster
A string, filename of cluster raster file

fn_imRrs.n
A string, filename of normalized Rrs raster file

fn_truecolorpng
A string, file name of truecolor png

fn_Chla
A string, file name of estimated Chla raster file

output_format
A string, the format of raster file, default as GTiff. See more in writeRaster

Value

A list() of all results and several inputs:

input What we input character link to image file or raster object

res Condition input of apply_to_image

raster.memb Raster object of membership value

raster.cluster Raster object of cluster

p.memb A ggplot list of membership value map

p.cluster A ggplot list of cluster map

p.truecolor A ggplot list of truecolor map

p.Chla A ggplot list of Chla map

imdf Data.frame of raw Rrs including xy coordinates

imRrs.raw Data.frame of raw Rrs

imRrs.n Data.frame of normalized Rrs

res.FCM List of FCM.new result

res.Chla Data.frame includes the template and final Chla estimation results

Details

The section FCM running used the subset of default Rrs clusters. Please see the section One more thing in vignettes New_Data_Running_FCMm if have not known how to do in that situation.

Also, if it is your first time to get the image data into R, you could load the raster data by typing raster::brick(filename) which filename is the path of your data such as '/data/Test.dat' or 'E://data//Test.tiff' or so.

References

Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.

Examples

if (FALSE) {
library(FCMm)
data("OLCI_TH")
data("Bi_clusters")
res <- generate_param(c(413,443,490,510,560,620,665,674,709,754,865,885))
im_result <- apply_to_image(input=OLCI_TH, res=res,
  title.name="Test_image", Chla_est=T, output_image=F)
}