Generate the input parameter of apply_to_image based on the built-in or user-defined centroids

Generate the input parameter of apply_to_image based on the built-in or user-defined centroids

generate_param(wl)

generate_param_ex(centroids, m = 1.36)

Arguments

wl
wavelength of subsetting

centroids
The centorids values required for fuzzy clustering on images. Its colnames should be converted to numeric wavelength. The format should be like Rrs_clusters.default.

m
The fuzzifier value of FCM. Default as 1.36.

Value

The return of generate_param or generate_param_ex is a list used for function apply_to_image.

Note

generate_param only support the cluster centroids proposed by Bi et al. (2019). However, the generate_param_ex could be used for user-defined centroids if you want.

References

Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.

See also

Examples

library(FCMm)
wl = c(413, 443, 490, 510, 560, 620, 665, 674, 681, 709, 754, 779, 865, 885)
res = generate_param(wl)