Estimate Chla concentration by algorithms blending via membership values

To calculate the Chla concentration via blending remote sensing algorithms.

FCM_m_Chla_estimation(Rrs, U)

Arguments

Rrs
Data.frame of Remote sensing reflectance which should contain colname with (also at) Rrs665, Rrs709, and Rrs754 with unit sr^-1.

U
Data.frame of membership values which should have seven columns presenting seven membership values from each cluster produced by FCM-m

Value

Each column presents Chla concentration with unit ug/L or mg/m^3 by model BR, TBA, C6 and blending result. Also with the membership values of each cluster.

Note

The input of Rrs must have bands with wavelength at 665, 709 and 754 nm. See examples of using this function.

References

Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.

Gilerson A A, Gitelson A A, Zhou J, et al. Algorithms for remote estimation of chlorophyll-a in coastal and inland waters using red and near infrared bands[J]. Optics Express, 2010, 18(23): 24109-24125.

See also

Examples

if (FALSE) {
library(FCMm)
data("WaterSpec35")
data("Bi_clusters")
Rrs <- WaterSpec35[,3:17]
result <- apply_FCM_m(Rrs=Rrs, option.plot=TRUE)
dt_Chla <- FCM_m_Chla_estimation(Rrs=data.frame(Rrs665=Rrs$`665`,
  Rrs709=Rrs$`708.75`,Rrs754=Rrs$`753.75`),U=result$u)
}