Apply FCM_m to new input Rrs

Application of FCM_m method for new Rrs data based on default cluster settings or user-defined clusters (trained by FCM.new).

apply_FCM_m(Rrs, wavelength, Rrs_clusters,
  stand=FALSE, default.cluster=TRUE, m_used=1.36, option.plot=FALSE)

Arguments

Rrs
Data.frame, the input Rrs of FCM.

wavelength
Vector, used for applying FCM. Default use the data from Bi_clusters.rda

Rrs_clusters
Data.frame, used for applying FCM. Default use the data from Bi_clusters.rda

stand
Logical, whether to normalized the Rrs data. Default as FALSE means do not.

default.cluster
Logical, whether to use the default clusters. Default use the data from Bi_clusters.rda

m_used
Number, Used fuzzifier value

option.plot
Logical, whether to plot the result. Default as FALSE

Value

A list including several results of function apply_FCM_m()

x The raw input Rrs dataframe with unit sr^-1

x.stand The standardized Rrs dataframe, if stand=F

d Distance to each cluster

u Membership values

cluster Defined by the maximum of membership

quality The quality of the cluster results.

m.used The used value of fuzzifier(m)

K Cluster number

p.group A ggplot list for plotting the cluster result

p.group.facet p.group with facet to see each cluster resutls more clearly

dt.plot Dataframe used for ggplot

References

Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.

Examples

if (FALSE) {
library(FCMm)
data("WaterSpec35")
data("Bi_clusters")
Rrs <- WaterSpec35[,3:17]
result <- apply_FCM_m(Rrs=Rrs, option.plot=TRUE)
}